backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
httpx
pydantic
pydantic-settings
python-dotenv
structlog
slowapi
pytest
pytest-asyncio